+++ Runtime/Serializers/ISerializer.cs
@@ -7,5 +7,6 @@ namespace ActionCode.Persistence
         string Serialize<T>(T data);
         string SerializePretty<T>(T data);
         T Deserialize<T>(string value);
+        void Deserialize<T>(string value, ref T objectToOverride);
     }
 }