+++ Runtime/Serializers/JsonUtilitySerializer.cs
@@ -6,18 +6,24 @@ namespace ActionCode.Persistence
     {
         public string Extension => "json";
 
+        public JsonUtilitySerializer()
+        {
+#if !MODULE_JSON_SERIALIZE
+            throw new System.Exception("Unity Json Module isn't installed. Install it using Package Manager.");
+#endif
+        }
+
+#if MODULE_JSON_SERIALIZE
         public string SerializePretty<T>(T data) => JsonUtility.ToJson(data, prettyPrint: true);
         public string Serialize<T>(T data) => JsonUtility.ToJson(data);
         public T Deserialize<T>(string value) => JsonUtility.FromJson<T>(value);
-
-        public static bool IsAvailable()
-        {
-#if MODULE_JSON_SERIALIZE
-            return true;
+        public void Deserialize<T>(string value, ref T objectToOverride) =>
+            JsonUtility.FromJsonOverwrite(value, objectToOverride);
 #else
-            Debug.LogError("Json Module isn't installed. Install it using Package Manager.");
-            return false;
+        public string SerializePretty<T>(T data) => default;
+        public string Serialize<T>(T data) => default;
+        public T Deserialize<T>(string value) => default;
+        public void Deserialize<T>(string value, ref T objectToOverride) { }
 #endif
-        }
     }
 }