+++ Runtime/Serializers/NewtonsoftSerializer.cs
@@ -8,24 +8,24 @@ namespace ActionCode.Persistence
     {
         public string Extension => "json";
 
+        public NewtonsoftSerializer()
+        {
+#if !PACKAGE_NEWTONSOFT
+            throw new System.Exception("Newtonsoft Json Module isn't installed. Install it using Package Manager.");
+#endif
+        }
+
 #if PACKAGE_NEWTONSOFT
         public string SerializePretty<T>(T data) => JsonConvert.SerializeObject(data, Formatting.Indented);
         public string Serialize<T>(T data) => JsonConvert.SerializeObject(data);
         public T Deserialize<T>(string value) => JsonConvert.DeserializeObject<T>(value);
+        public void Deserialize<T>(string value, ref T objectToOverride) =>
+            JsonConvert.PopulateObject(value, objectToOverride);
 #else
         public string SerializePretty<T>(T data) => Serialize(data);
         public string Serialize<T>(T _) => string.Empty;
         public T Deserialize<T>(string _) => default;
+        public void Deserialize<T>(string value, ref T objectToOverride) { }
 #endif
-
-        public static bool IsAvailable()
-        {
-#if PACKAGE_NEWTONSOFT
-            return true;
-#else
-            UnityEngine.Debug.LogError("Newtonsoft Json package isn't installed. Install it using Package Manager.");
-            return false;
-#endif
-        }
     }
 }